Introduced local variable to not overwrite pee
The issue with several <pre> tags stems from the variable being
overwritten every time the loop iterated, which explains why only the
last <pre> tag persisted in the output. This wasn't a problem for most
test cases as they only made sure 1 <pre> tag behaved like it was
meant to.